Rewrite logic for positioning popup
As I've been unable to find out exactly what's triggered #2 I figured
I'd just try to rewrite the logic for positioning the popup and hope the
problem magically disappeared.  It almost did: the only situations where
I experienced issues after the rewrite was when I transitioned from a
wide completion candidate popup to a small one. This would cause the
quickhelp popup to stay at the previous position and be misaligned.
After setting `company-tooltip-minimum-width` to 40 the difference
between the very large completion popups and the small ones are so small
that the problem disappeared. It's unfortunate that we have to override
a user setting like this, but having a misaligned help popup is
certainly more annoying.