[Fix #8] Suppress completing reading
CIDER, and probably other libs, prompt the user to resolve ambiguities
in documentation requests.  When these requests are made by
company-quickhelp such popups can be incredibly frustrating.

Instead of just throwing an error, indicating that we don't know which
candidate to present documentation for, I've opted to just grab the
first candidate.  This is slightly more useful than just erroring out,
even if it can be argued that it isn't the 'right thing' to do.