[Fix #16] Delay checking for compatibility
We can't check if the user is running company-quickhelp in the terminal
when the mode is turned on because it might be turned on as part of
initializing emacs in daemon mode.  At that point emacs is indeed
running in the terminal but it's also not problematic.

The compat check is now delayed until the user actually attempts to use
company-quickhelp to create a popup.  At that point we error out.

This solution isn't perfect because the user is likely to see quite a
few errors as she navigates around the completion list, but I worry that
if I just display the error once, and then disable the mode, the user is
going to miss it error message and wonder why this piece of shit
software isn't working as advertised.