Prevent constant errors when enabled in terminal
It's possible for company-quickhelp to be enabled in terminal Emacs
frames, e.g. by starting a GUI frame and then opening additional
terminal frames with emacsclient. And, indeed, it should logically
always be safe to enable company-quickhelp-mode in any buffer, because
that buffer may be displayed in terminal or GUI frames at any future
time.

Without this change, when company calls the company-quickhelp backend
in terminal frames, pos-tip-hide is called unconditionally, which
results in errors. For real-world examples, see
https://github.com/purcell/emacs.d/issues/377 and
https://github.com/purcell/emacs.d/issues/414.

This commit adds a function called
company-quickhelp-pos-tip-available-p, which is used to detect and
prevent these errors.