Don't error when pos-tip is unavailable
The result of `company-quickhelp-pos-tip-available-p` can vary from frame to frame, since even on a system that supports `x-hide-tip` etc., `window-system` will be different between frames: in terminal frames `window-system` will be nil.

And, in fact, the same buffer with `company-quickhelp-local-mode` enabled  can be visible in two different frames simultaneously. The existing code will trigger a hard error in one of those frames, but not the other. The behaviour of `company-quickhelp--hide` is the correct one here: silently do nothing when quickhelp is unavailable. This commit changes the behaviour of `company-quickhelp--show` to match.

See also the resulting error https://github.com/purcell/emacs.d/issues/554